upload-images/django_test_project/django_test_project/requirements.txt
asn1crypto
backports.ssl-match-hostname
cached-property
certifi
chardet
cloudinary
cryptography
Django
djongo
django-cors-headers
djangorestframework
django-braces
docker
docker-compose
docker-pycreds
dockerpty
pymongo
urllib3
python-consul
requests-mock
pytest-mongodb
pytest 
